Remove alert service helper
